add search pagination test
